Refactor load_ordered_yaml into new yamlutils mod
This allows load_ordered_yaml to be used by both Specification and
Metric YAML parsing code.
